server/requirements.txt
pyhumps
fastapi
uvicorn
pytest
fastapi-sqlalchemy
pydantic
psycopg2
pandas
sqlalchemy_utils